Move utils into internal package
